requirements.txt
bs4
coverage
feedparser
html5lib
ignition-gemini
service_identity
twisted
websockets
yahoo_fin
